fix: background update for secret20 viewing permission

+++ packages/fetch-extension/src/pages-new/sign/eth-sign.tsx
@@ -316,7 +316,7 @@ export const SignEthereumPage: FunctionComponent = observer(() => {
                             setApproveButtonClicked(true);
                             signature = await handleEthereumPreSignByLedger(
                               interactionData,
-                              signingDataBuff,
+                              signingDataBuff as Uint8Array,
                               {
                                 useWebHID: ledgerInitStore.isWebHID,
                               }
